Fixed issue with scaling algorithm
Fixed an issue with the scaling algorithm. The original algorithm didn't
take into account balancing of VMs across Fault Domains. In theory, you
could end up with a VMSS with many VMs, but all VMs in the same Fault
Domain.
The new algorithm first derermines which Fault Domain has the most
active VMs, and elects the one with the lowest load to drain.